requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
python-dotenv==1.0.0
structlog==23.3.0
psycopg2==2.9.10
